requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
pandas>=2.2.0
markdown-pdf>=1.3.0
python-dotenv>=1.0.0
